- Fixed NNTP server 100% CPU usage - Implemented NNTP server authentication - Enabled NNTP server, it works now

+++ src/plugins/Freetalk/Freetalk.java
@@ -159,7 +159,7 @@ public class Freetalk implements FredPlugin, FredPluginFCP, FredPluginL10n, Fred
 		
 		// FIXME: Fix this: https://bugs.freenetproject.org/view.php?id=2977
 		//Logger.debug(this, "Starting NNTP server...");
-		//mNNTPServer = new FreetalkNNTPServer(mPluginRespirator.getNode(), this, 1199, "127.0.0.1", "127.0.0.1");
+		mNNTPServer = new FreetalkNNTPServer(mPluginRespirator.getNode(), this, 1199, "127.0.0.1", "127.0.0.1");
 		//mNNTPServer = new FreetalkNNTPServer(mPluginRespirator.getNode(), this, 1199, "0.0.0.0", "*");
 		
 		Logger.debug(this, "Plugin loaded.");
@@ -275,11 +275,13 @@ public class Freetalk implements FredPlugin, FredPluginFCP, FredPluginL10n, Fred
 			Logger.error(this, "Error during termination.", e);
 		}
 		
-		try {
-			mNNTPServer.terminate();
-		}
-		catch(Exception e) {
-			Logger.error(this, "Error during termination.", e);
+		if (mNNTPServer != null) {
+    		try {
+		        mNNTPServer.terminate();
+    		}
+    		catch(Exception e) {
+    			Logger.error(this, "Error during termination.", e);
+    		}
 		}
 		
         try {
